+++ src/s21_math.h
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#define s21_NAN (0.0 / 0.0L)
-#define s21_INF (1.0 / 0.0L)
+#define s21_NAN (0.0 / 0.0)
+#define s21_INF (1.0 / 0.0)
 #define s21_TOLERANCE 1e-6
 #define s21_PI 3.14159265358979323846
 
@@ -12,7 +12,7 @@ int compare_float(double f1, double f2);
 long double s21_cos(double x);
 
 // Computes sine
-double s21_sin(double x);
+long double s21_sin(double x);
 
 // Computes absolute value of an integer value
 int s21_abs(int x);
